refactoring report
